Attention: All settings are lost
I have modified the the datatypes and interpretation of the settings and did ditch the backward compatibility in favor of better readability.

Please try out $1 for a guided input setup.

+++ inc/sbus.h
@@ -65,7 +65,7 @@ extern sbusData_t sbusdata;
 
 void SBUS_IRQHandler(UART_HandleTypeDef *huart);
 void printSBUSChannels(Endpoints endpoint);
-int16_t getDuty(uint8_t channel);
+float getDuty(uint8_t channel);
 uint8_t* getSBUSFrameAddress(void);
 void HAL_TIM_IC_CaptureCallback(TIM_HandleTypeDef *htim);
 void initSBusData(uint8_t receivertype);
